requirements.txt
# PDF 印刷增强工具 — 依赖清单
# 安装：pip install -r requirements.txt

# ── 核心依赖 ──────────────────────────────────────────────
PyMuPDF>=1.27.0          # PDF 解析、页面重建、图像提取
Pillow>=10.0.0           # 图像处理、ICC 色彩转换
pikepdf>=10.0.0          # 真 CMYK 字节流后处理（v2.1 新增）
numpy>=1.24.0            # 图像增强数值计算

# ── 可选 GUI 依赖 ────────────────────────────────────────
# tkinterdnd2>=0.4.0    # GUI 拖放支持（Windows/macOS，可选；未装则降级为按钮选择）

# ── 验证与调试依赖 ────────────────────────────────────────
rich>=13.0.0             # validate_pipeline.py 彩色终端输出
pytest>=9.0.0            # 单元测试（tests/ 目录）
opencv-python-headless>=4.8.0  # CLAHE 局部对比度增强（CV2）

# ── 可选 AI 依赖 ─────────────────────────────────────────
# Real-ESRGAN             # AI 超分（需手动下载模型，非 pip 安装）
# （硅基流动视觉模型通过环境变量 SILICONFLOW_API_KEY 接入，无需 pip 包）
